add message

+++ Sources/ATTT/PH2/GUI/TRUONGDV/TruongDV_XemPC.cs
@@ -79,22 +79,21 @@ namespace PH2.GUI.TRUONGDV
             tmp = dataGridView1.Rows[rowIndex].Cells[5].Value.ToString();
             int NAM = Int32.Parse(tmp);
             string MACT = dataGridView1.Rows[rowIndex].Cells[6].Value.ToString();
-            try
+            DialogResult dialogResult = MessageBox.Show("Dữ liệu của lớp giảng dạy có thể bị mất!", "Thông báo", MessageBoxButtons.YesNo);
+            if (dialogResult == DialogResult.Yes)
             {
-                XemPCBLL.DeleteRow(MAGV, MAHP, HK, NAM, MACT);
+                try
+                {
+                    XemPCBLL.DeleteRow(MAGV, MAHP, HK, NAM, MACT);
+                }
+                catch (Exception ex)
+                {
+                    MessageBox.Show("Lỗi xóa khóa ngoại!");
+                    return;
+                }
+                Load();
+                MessageBox.Show("Xóa phân công thành công!");
             }
-            catch (Exception ex)
-            {
-                MessageBox.Show("Lỗi xóa khóa ngoại!");
-                return;
-            }
-            Load();
-            MessageBox.Show("Xóa phân công thành công!");
-        }
-
-        private void label3_Click(object sender, EventArgs e)
-        {
-
         }
 
         private void addButton_Click(object sender, EventArgs e)
@@ -120,13 +119,16 @@ namespace PH2.GUI.TRUONGDV
                 if (ex.Message.Contains("policy"))
                 {
                     MessageBox.Show("Vi phạm chính sách VPD!");
-                    return;
                 }
                 else if (ex.Message.Contains("unique constraint"))
                 {
-                    MessageBox.Show("Kế hoạch mở không có học phần này hoặc đã được đăng ký!");
-                    return;
+                    MessageBox.Show("Kế hoạch đã được đăng ký!");
                 }
+                else
+                {
+                    MessageBox.Show("Kế hoạch mở không có học phần này!");
+                }
+                return;
             }
             MessageBox.Show("Thêm vào thành công!");
             Load();
